Ueransim (#7)
* PR: Added ueransim component

+++ .global/pac/DEPLOY_COMPONENT.groovy
@@ -9,33 +9,36 @@ pipeline {
 
   parameters {
     string(name: 'TN_ID', defaultValue: '', description: 'TRIAL NETWORK IDENTIFIER')
-    string(name: 'LIBRARY_COMPONENT_NAME', defaultValue: '', description: '6G LIBRARY COMPONENT')
+    string(name: 'LIBRARY_URL', defaultValue: 'https://github.com/6G-SANDBOX/6G-Library', description: '6G LIBRARY REPOSITORY')
     string(name: 'LIBRARY_BRANCH', defaultValue: 'main', description: '6G LIBRARY BRANCH')
+    string(name: 'LIBRARY_COMPONENT_NAME', defaultValue: '', description: '6G LIBRARY COMPONENT')
     string(name: 'DEPLOYMENT_SITE', defaultValue: 'uma', description: 'Site where deploy')
     base64File (name: 'FILE', description: 'YAML file that contains variables needed to deploy the component')
+    booleanParam(name: 'DEBUG', defaultValue: false, description: 'Enable DEBUG')
     //string(name: 'TNLCM_CALLBACK', defaultValue: 'https://tnlcm.uma/TN/ID/callback', description: 'URL of the TNLCM to notify result')   
   }
 
   environment {
-      LIBRARY_URL="https://github.com/6G-SANDBOX/6G-Library"
       TN_ID="${params.TN_ID}"
-      LIBRARY_COMPONENT_NAME="${params.LIBRARY_COMPONENT_NAME}"
+      LIBRARY_URL="${params.LIBRARY_URL}"
       LIBRARY_BRANCH="${params.LIBRARY_BRANCH}"
+      LIBRARY_COMPONENT_NAME="${params.LIBRARY_COMPONENT_NAME}"
       DEPLOYMENT_SITE="${params.DEPLOYMENT_SITE}"
       FILE_PATH="${WORKSPACE}/${params.LIBRARY_COMPONENT_NAME}/private/tnlcm_vars.yaml"
       OPENNEBULA_API_CREDENTIALS = credentials('OPENNEBULA_API_CREDENTIALS')
-      OPENNEBULA_USERNAME = credentials('OPENNEBULA_USERNAME')
-      OPENNEBULA_PASSWORD = credentials('OPENNEBULA_PASSWORD')
+      OPENNEBULA_USERNAME = credentials('OPENNEBULA_TNLCM_USERNAME')
+      OPENNEBULA_PASSWORD = credentials('OPENNEBULA_TNLCM_PASSWORD')
       OPENNEBULA_ENDPOINT = credentials('OPENNEBULA_ENDPOINT')
+      OPENNEBULA_FLOW_ENDPOINT = credentials('OPENNEBULA_FLOW_ENDPOINT')
       OPENNEBULA_INSECURE = credentials('OPENNEBULA_INSECURE')
       // MINIO CREDENTIALS
       AWS_ACCESS_KEY_ID = credentials('MINIO_KEY')
       AWS_SECRET_ACCESS_KEY = credentials('MINIO_SECRET')
-      //ANSIBLE
+      // ANSIBLE
       ANSIBLE_REMOTE_USER = credentials('ANSIBLE_REMOTE_USER')
       ANSIBLE_CONNECTION_PASSWORD_FILE = credentials('ANSIBLE_CONNECTION_PASSWORD_FILE')
-      //ANSIBLE_BECOME_PASSWORD_FILE = credentials('ANSIBLE_BECOME_PASSWORD_FILE')    
-      //GITHUB JENKINS TOKEN
+      // ANSIBLE_BECOME_PASSWORD_FILE = credentials('ANSIBLE_BECOME_PASSWORD_FILE')    
+      // GITHUB JENKINS TOKEN
       GITHUB_JENKINS = credentials('GITHUB_JENKINS')  
   }
 
@@ -48,7 +51,6 @@ pipeline {
                 script: 'echo $FILE | base64 -d',
                 returnStdout: true)
             echo "${YAML_CONTENT}"
-            // This file path is relevant because ansible will look for this file on stage 2
             writeFile(file: FILE_PATH, text: YAML_CONTENT)
           }
       }
@@ -62,7 +64,7 @@ pipeline {
               """
           }
       }
-    }      
+    }
     stage('Stage 2: Run component deployment') {
       steps {
         script {
@@ -82,15 +84,22 @@ pipeline {
           '''
       }
       cleanup{
-          /* clean up our workspace */
-          deleteDir()
-          /* clean up tmp directory */
-          dir("${env.workspace}@tmp") {
-              deleteDir()
-          }
-          /* clean up script directory */
-          dir("${env.workspace}@script") {
-              deleteDir()
+          script {
+              if (env.DEBUG == 'false') {
+                  echo 'CleanUp Workspace'
+                  /* clean up our workspace */
+                  deleteDir()
+                  /* clean up tmp directory */
+                  dir("${env.workspace}@tmp") {
+                      deleteDir()
+                  }
+                  /* clean up script directory */
+                  dir("${env.workspace}@script") {
+                      deleteDir()
+                  }    
+              } else {
+                  echo 'Workspace is not deleted'
+              }
           }
       }      
   }
